refactor individual scripts
